Easy throttle changes
This is an experimental code but very important milestone. The nature of PWM in Arduino is to provide continuous pulse with specific duty cycle. I don't know this is an effect of servo library or the way timer-hardware for PWM in Arduino works. But it made it easier to get the car moving.